packages/schellenberghack/pyproject.toml
[project]
name = "schellenberghack"
version = "0.9.0"
dependencies = ["pyserial>=3.5"]

[build-system]
requires = ["uv_build>=0.9.29,<0.10.0"]
build-backend = "uv_build"
